+++ BeyondTrust/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2026-05-22 - 1.3.1
+
+### Changed
+
+- Add scalability labels to Prometheus metrics
+
 ## 2026-04-30 - 1.3.0
 
 ### Added