+++ Bitsight/CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 2026-07-30 - 1.3.0
+## 2026-05-22 - 1.2.2
 
 ### Changed
 
-- Migrated module from Poetry to uv
+- Add scalability labels to Prometheus metrics
 
 ## 2025-06-16 - 1.2.0